fix: resolve Next.js build and routing conflicts
- Split blog page into server/client components (generateMetadata conflict)
- Remove static sitemap.xml (route specificity conflict with sitemap.ts)

Fixes deployment build failures.
